+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add the option to rebuild even if it's not needed
+- Add the option to speciry the path to the directory
 
 ### Changed
 - Regex variables are now weaker than local ones