+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.15.0] - 2020-05-07
 ### Added
 - Add the option to rebuild even if it's not needed
 - Add the option to speciry the path to the directory